# viz: non blocking UOp tracing