--- a/tinygrad/uop/ops.py
+++ b/tinygrad/uop/ops.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import Any, Optional, Union, Callable, cast, TYPE_CHECKING, Type, Sequence
 import sys, time, functools, itertools, math, operator, hashlib, os, types, pickle, pathlib, inspect, weakref
-from dataclasses import dataclass, field
+from dataclasses import dataclass, field, replace
 from tinygrad.uop import Ops, GroupOp
 from tinygrad.uop.mathtraits import MathTrait
 from tinygrad.dtype import ConstType, ImageDType, dtypes, DType, truncate
@@ -79,6 +79,7 @@ class UOp(MathTrait, metaclass=UOpMetaClass):
       if (ref:=UOpMetaClass.ucache.get(k:=(self.op, self.dtype, self.src, self.arg, self.tag))) is not None:
         for s in self.src: s.children.discard(ref)
         del UOpMetaClass.ucache[k]
+        uop_number.pop(ref, None)
     except AttributeError: pass
   def __reduce__(self):
     args = [self.op, self.dtype, self.src, self.arg, self.tag, self.metadata]
@@ -719,6 +720,19 @@ class PatternMatcher:
     while new_n is not None: last_n, new_n = new_n, self.rewrite(new_n, ctx)
     return last_n
 
+# *** non-blocking UOp tracker ***
+
+ucount = itertools.count()
+uop_number:dict[weakref.ReferenceType[UOp], int] = {}
+uop_fields:dict[int, tuple] = {}
+def track_uop(u:UOp) -> int:
+  if (cret:=uop_number.get(ref:=weakref.ref(u))) is not None: return cret
+  uop_number[ref] = num = next(ucount)
+  # KERNEL also has UOp in the arg, TODO: fix this to be generic.
+  arg = replace(arg, ast=track_uop(arg.ast)) u.arg if u.op is Ops.KERNEL else u.arg
+  uop_fields[num] = (u.op, u.dtype, tuple(track_uop(s) for s in u.src), arg, u.tag)
+  return num
+
 # *** tracking pattern matcher ***
 
 VIZ = ContextVar("VIZ", 0)
@@ -727,8 +741,8 @@ match_stats:dict[UPat, list[Union[int, float]]] = dict()
 @dataclass(frozen=True)
 class TrackedGraphRewrite:
   loc: tuple[str, int]                                                                       # location that called graph_rewrite
-  sink: UOp                                                                                  # the sink input to graph_rewrite
-  matches: list[tuple[UOp, UOp, UPat]]                                                       # before+after of all the matches
+  sink: int                                                                                  # the sink input to graph_rewrite
+  matches: list[tuple[int, int, UPat]]                                                       # before+after of all the matches
   name: str|None                                                                             # optional name of the rewrite
   depth: int                                                                                 # depth if it's a subrewrite
   bottom_up: bool
@@ -770,7 +784,7 @@ def track_matches(func):
     if tracking:=(TRACK_MATCH_STATS >= 2 and tracked_ctxs):
       loc = ((frm:=sys._getframe(1)).f_code.co_filename, frm.f_lineno)
       depth = len(active_rewrites)
-      tracked_ctxs[-1].append(ctx:=TrackedGraphRewrite(loc, args[0], [], kwargs.get("name", None), depth, kwargs.get("bottom_up", False)))
+      tracked_ctxs[-1].append(ctx:=TrackedGraphRewrite(loc, track_uop(args[0]), [], kwargs.get("name", None), depth, kwargs.get("bottom_up", False)))
       active_rewrites.append(ctx)
     ret = func(*args, **kwargs)
     if tracking: active_rewrites.pop()
@@ -792,7 +806,7 @@ class TrackedPatternMatcher(PatternMatcher):
         match_stats[p][0] += 1
         match_stats[p][3] += (et:=time.perf_counter()-st)
         if TRACK_MATCH_STATS >= 3: print(f"{et*1e6:7.2f} us -- ", p.printable())
-        if TRACK_MATCH_STATS >= 2 and isinstance(ret, UOp) and active_rewrites: active_rewrites[-1].matches.append((uop, ret, p))
+        if TRACK_MATCH_STATS >= 2 and isinstance(ret, UOp) and active_rewrites: active_rewrites[-1].matches.append((track_uop(uop),track_uop(ret), p))
         return ret
       match_stats[p][2] += time.perf_counter()-st
     return None
@@ -805,7 +819,7 @@ if TRACK_MATCH_STATS or PROFILE:
     if TRACK_MATCH_STATS >= 2:
       with open(fn:=temp("rewrites.pkl", append_user=True), "wb") as f:
         print(f"rewrote {len(tracked_ctxs)} graphs and matched {sum(len(r.matches) for x in tracked_ctxs for r in x)} times, saved to {fn}")
-        with Context(PICKLE_BUFFERS=0): pickle.dump((tracked_keys, tracked_ctxs), f)
+        with Context(PICKLE_BUFFERS=0): pickle.dump((tracked_keys, tracked_ctxs, uop_fields), f)
     if VIZ: launch_viz("VIZ", temp("rewrites.pkl", append_user=True))
     if getenv("PRINT_MATCH_STATS", 1):
       ret = [0,0,0.0,0.0]