--- a/tinygrad/uop/ops.py
+++ b/tinygrad/uop/ops.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import Any, Optional, Union, Callable, cast, TYPE_CHECKING, Type, Sequence
 import sys, time, functools, itertools, math, operator, hashlib, os, types, pickle, pathlib, inspect, weakref
-from dataclasses import dataclass, field, replace
+from dataclasses import dataclass, field, fields, is_dataclass, replace
 from tinygrad.uop import Ops, GroupOp
 from tinygrad.uop.mathtraits import MathTrait
 from tinygrad.dtype import ConstType, ImageDType, dtypes, DType, truncate
@@ -722,16 +722,20 @@ class PatternMatcher:
 
 # *** non-blocking UOp tracker ***
 
+@dataclass(frozen=True)
+class UNum: n:int
 ucount = itertools.count()
-uop_number:dict[weakref.ReferenceType[UOp], int] = {}
-uop_fields:dict[int, tuple] = {}
-def track_uop(u:UOp) -> int:
-  if (cret:=uop_number.get(ref:=weakref.ref(u))) is not None: return cret
-  uop_number[ref] = num = next(ucount)
-  # KERNEL also has UOp in the arg, TODO: fix this to be generic.
-  arg = replace(u.arg, ast=track_uop(u.arg.ast)) if u.op is Ops.KERNEL else u.arg
-  uop_fields[num] = (u.op, u.dtype, tuple(track_uop(s) for s in u.src), arg, u.tag)
-  return num
+uop_number:dict[weakref.ReferenceType[UOp], UNum] = {}
+uop_fields:dict[UNum, tuple] = {}
+def track_uop(a:Any):
+  if isinstance(a, UOp):
+    if (cret:=uop_number.get(ref:=weakref.ref(a))) is not None: return cret
+    uop_number[ref] = num = UNum(next(ucount))
+    uop_fields[num] = (a.op, a.dtype, track_uop(a.src), track_uop(a.arg), a.tag)
+    return num
+  if isinstance(a, tuple): return type(a)(track_uop(i) for i in a)
+  if is_dataclass(a): return replace(a, **{f.name:track_uop(getattr(a, f.name)) for f in fields(a)})
+  return a
 
 # *** tracking pattern matcher ***