The core idea is to track UOp fields, and lazily reconstruct UOp instances when the user opens a graph.
Benchmarks:

`VIZ=1 PYTHONPATH=. python3 examples/llama3.py --shard 4 --benchmark`
Master:  2.3 tok/s
Branch: 63.20 tok/s
VIZ = 0: 87.94 tok/s

This also naturally improves VIZ startup time, because it doesn't have UOp unpickling overhead on startup:

llama: 5800 ms → 3150 ms
beautiful_mnist: 440 ms → 215 ms

The fix for KERNEL UOp should be made generic, currently some UOps can have other UOps in the arg.